Enable the dT_crit functionality for the VODE burner
This reproduces a piece of code that was in some of the burners
we used before the switch to Microphysics. dT_crit is a real
variable that expresses the fractional change in temperature
one would like between calls to the EOS to re-evaluate the
thermodynamic derivatives (i.e. c_v and c_p). One new thing is
that in between EOS calls we do a first-order prediction of those
specific heats using a numerical difference estimate of dc_v/dt
and dc_p/dt.