get this working + using nspec_evolve
previously, it was not considering the ash update at all.  Now
we switch to nspec_evolve = 1, and algebraically compute the
update to the ash from the mass fractions.

The energy output from this now agrees very well with Maestro,
but the ash mass fraction is off by 2x -- not sure what's up
yet.