Fix bug: min, not max, in mazurek
There was a bug in the mazurek subroutine for the electron capture
rates on nickel 56. This had the effect of resetting any temperature
below the maximum for that routine (T9 = 14) to the maximum. The fix
replaces a max with a min so that any temperature in the allowed range
of (2 < T9 < 14) is correctly fit, and any temperature above T9 = 14
is reset to T9 = 14. This has been corrected in the main releases of
the aprox* networks as well as MESA.