Do not rely on the numerical Jacobian for aprox19/aprox21 (#49)
* Do not rely on the numerical Jacobian for aprox19/aprox21

The change was made at some point to use the numerical Jacobian for
these networks, because we learned that the hardwired Jacobian has
some missing terms. However, the numerical Jacobian in practice seems
to have big problems -- it can cause VODE to complete the integration
with crazy states that have X >> 1, but VODE thinks it is converged.
We also see NaN's in the Jacobian when we examine it. So we are going
to revert to the analytical Jacobian default -- this seems to work better.

* Update CHANGES for aprox Jacobian changes