Removed yhscratch copy of Nordsieck array.
This was previously used to multiply the Nordsieck array by the Pascal
triangle matrix in the same 1-D loop fashion as the original VODE.

Now, we implement this array multiplication via additions in 2-D loops
through the Nordsieck array, obviating the need for an array copy.

Subroutine `advance_nordsieck` is used in forming the predicted value
of the Nordsieck array, and subroutine `retract_nordsieck` is used to
undo the prediction in case the step fails.