For simplified SDC, apply advection sources at end of integration using integration time (#213)
When calling vode_to_sdc at the end of the simplified SDC integration with VODE and VODE90, we were previously passing time as the time step to update unevolved variables rho and rho*u.

time was the simulation time in Castro at the start of the burner call, not the ODE integration end time. The difference is that the ODE integration end time is just the time advanced by the ODE integrator relative to its starting time of 0. So this is the reaction timestep from Castro that we should use to apply the advection sources to rho and rho*u before returning to Castro.

Note: Castro was not using the result of this last update, so no changes are expected.