get rid of the xin files for test_react (#211)
This does away with the composition input files xin.* in favor of computing the composition on the fly. The advantage of this is that we can easily change the size of the test to investigate scaling. There are now 3 runtime parameters (to be set in the &extern namelist): primary_species_1, primary_species_2, primary_species_3. These primary species will vary from X = 0.2 to X = 0.7 to 0.9 (depending on the number). Only one primary species varies at a time. The non-primary species will be set equally to share whatever fraction of 1 is not accounted for by the primary species mass fractions.

Note: tests will be completely different because we are completely changing the species info.