add nuclei into to the C++ header (#244)
This renames the actual_network.H file to network_properties.H and updates it to include aion and zion. A .net file is present in each network now that creates the nuclei information for both Fortran and C++. This is done by running the top-level update_headers.py script. That writes actual_network.H and a new module, network_parameters.F90. The actual_network.F90 has been rewritten to use network_parameters.F90.