Convert aprox19 to C++ (#320)
I've tested this on both test_rhs and test_react. For test_rhs, the maximum relative difference between the Fortran and C++ is O(1e-15) for the RHS, and O(1e-13) for the Jacobian. For test_react, the maximum relative difference is 1e-11 for the CPU build and 1e-4 for the GPU build. (The difference for the GPU build is of comparable magnitude to what we see for the other networks.)