Remove array over NumGroups from rate_t (#416)
In every place in the networks where we are using rate_t, it works just as well to have an array of rate_t, where the size of the array is the number of rate groups, as it does to have the rate_t be a 2D array. Making this change has two benefits: we can clean up some of the RHS code later, because the RHS (unlike the Jacobian) typically does not actually need to fill all the rate groups (so this will save work and memory), and also this will support the change in #413.