Add interface for C++ integrator to use Fortran RHS (#419)
If a network has a C++ RHS implementation, it should set DEFINES += -DNETWORK_HAS_CXX_IMPLEMENTATION in its Make.package, and the C++ integrator will continue to use it. Otherwise, we default to the Fortran RHS, and we add a wrapper script that translates between the C++ RHS/Jac and the Fortran RHS/Jac.

With this change, any downstream application can use USE_CXX_REACTIONS = TRUE everywhere and still use any network. The only catch is that there is no CUDA implementation right now in this PR, so the legacy Fortran networks won't run on GPUs. If we need to fix that later, it should be possible.