Implement iso7 RHS terms in a standardized form using data from a constexpr function (#413)
This PR is the start of a series of refactors so that iso7's RHS is re-implemented by moving all of the RHS data to the header in a function called rhs_data that returns an rhs_t struct. Every term in an RHS is a multiplicative combination of up to five terms: a prefactor, Y(i), Y(j), Y(k), and rate(l). Most reactions have two Y factors, like C12 + C12 -> Mg24; some have only one Y factor, like C12 + gamma -> 3He4; and the three-body reactions (triple alpha) can have three. If the species index is -1, we know to ignore it. If the compiler can optimize this suitably well by evaluating these terms at compile time, this should add no additional runtime cost. This initial change will do the evaluation at run time and incur a performance cost, but that will be resolved in a future PR.

The motivation for this change is that we want the flexibility to experiment with integrators that can directly utilize information about the form of the RHS. Right now the RHS is a black box to the integrator, but if we implement every RHS in a standard form like this, then we can write general integrators that can perform more sophisticated integration algorithms (but that can still construct the standard form of the RHS if needed, since they have all the terms ready to do so).

From the implementation perspective, I'm constructing the RHS by looping over rates, identifying the rates that contribute to the RHS for each species, and then filling the array which we then pass to esum.

With this infrastructure in place, and because we now fully construct the RHS behind the scenes, we should be able to do any number of things to experiment with integration schemes. For example, we could look into the idea I mentioned of constructing the RHS as a running esum rather than explicitly constructing the array and only then applying esum to it. We could also experiment with constructing the rates one by one and filling the RHS as we go, rather than explicitly constructing the full rate array. More generally I want to experiment with other integration strategies which require access to direct information about the structure of the RHS terms.

For this first PR, I've only applied it to one term in iso7 -- the Ni56 RHS -- and for now we're doing the evaluation at run time, not compile time. This will temporarily slow down iso7 but I have been able to demonstrate that when the changes are done, the performance will be recovered.