iso7 part 8 (#449)
iso7 has a rate dependence on Y since the (7*He4 + Si28 -> Ni56) effective rate (both forward and reverse) depends on Y. The forward rate is given by k(rho, T) * Y(He)^3 * Y(Si28), so several Jacobian elements need to have d(rate)/dY terms to be computed correctly. (A similar thing is true for the reverse rate.) The current code knows exactly which d(rate)/dY terms are computed and stores them in rate arrays dratedy1 and dratedy2. But if we want to generalize this we need to be agnostic to which specific rates and which specific species have the dependence. A solution is to allow the Jacobian to generally be a function of the derivatives of the rates with respect to Y in addition to the derivatives of Y itself in the RHS. To effect this generally we need to have a dratedY container which has NumSpec species and NumRates rates, and this change enables that. The vast majority of these entries will be empty, so this temporarily increases the storage requirements and adds some redundant calculations. But the change we'll be doing later is to detect the nonzero ones and only add them in at compile time.