iso7 9/∞ (#455)
Now we start applying the same approach to the Jacobian that we did for the RHS. We assert that the Jacobian terms have the same form as the RHS terms, K * rate * Y1 * Y2 * Y3 (although I think there cannot be any three-body terms). A complicating factor is that the rate derivative d(rate) / dY may be non-zero since many of the rates in the aprox networks have hidden contributions from equilibrium sequences that depend on Y. So in addition the Jacobian have a term from K * drate/dY * Y1 * Y2 * Y3.