Add explicit integrator with asymptotic approximation (#466)
The explicit integration method of Guidry et al. (2013) (paper 1 of 3) is implemented. In this approach we detect which terms in the species RHS would equilibrate within the current integration timestep and then use the asymptotic approximation to update that term instead of normal explicit integration.

This seems to work OK for "easy" burns at low temperature -- the answer is fairly inaccurate for the default choice of parameters, presumably due to the first order method, but if we are very restrictive on the timestep then the answer converges to what VODE provides. Near NSE we just run into the same problems as the simple forward Euler integrator, which will require the "partial equilibrium" method that detects which specific rates are in approximate detailed balance and removes them from the integration.

To do:

* Use a higher-order integration method for both the explicit and asymptotic terms (the paper describes a predictor-corrector scheme)
￼ * Add the partial equilibrium method from paper 3 of that series