iso7 14/∞ (#467)
Evaluate as much of the RHS as possible at compile time. Specifically we do not waste memory or flops on contributions to each RHS from rates that don't actually contribute.

To make this simpler to review I am retaining the old behavior that the esum is over NumRates terms even though most of them are zero. I will do the piece that skips the zero terms in a later PR.