Add types to determine if an EOS type has given members (#481)
To do #470 we will need to be able to determine at compile time if a given template type has certain member data. This adds a few structs for that purpose. The idea is that they can be used like

    if constexpr (has_xne_xnp<T>::value) {
        state.xne = xnefer;
        state.xnp = 0.0e0_rt;
    }

The intended usage above requires C++17, but this PR does not yet require that.