remove the BS integrator (#488)
BS was Fortran only, so we couldn't use it in our codes anyway
it also was not very good. We also now have a C++ ForwardEuler if
we want to generate a reference solution (with really small timesteps)
for testing. This change also removes the integrator switch logic
in the Fortran driver.