Only accesses variables in the state in Helmholtz (#507)
Continues #470 by compiling out accesses to members of the struct sent to the EOS that are invalid. This will allow us later to, for example, send a type to the EOS that doesn't have entropy or enthalpy.

We don't explicitly compile out all intermediate expressions that led up to the quantities in question. That could be done, but compilers should be able to recognize that as dead code and compile it out for us.