start reusing some code for runtime parameters (#510)
This creates a centralized Params class in runtime_paramerters.py that eventually will be used by the application codes and Microphysics for managing runtime parameters. It has all of the necessary methods to write the C++ and Fortran needed for when Fortran manages the runtime parameters (as currently is the case in Microphysics) and for when C++ manages them (like we do in Castro for the main parameters).