Start moving towards zero-based indexing with linpack (#536)
RArray1D and RArray2D are one-based arrays, and that causes friction with the zero-based indexing of xn and other arrays. So we want to move everything to zero-based indexing.

The strategy will be to suffix function parameters with "1", e.g. a -> a1, and then do

auto& a = reinterpret_cast<Array2D<Real, 0, num_eqs-1, 0, num_eqs-1>&>(a1)

Then, in the function, we can convert all the calculations to zero-based indexing. When all of the conversions have been done, we can do one final sweep to convert RArray1D and RArray2D to be zero-based and then remove all the reinterpret_cast statements and rename the function parameters to their original names. This will allow us to do an incremental conversion.