SDC+NSE part II (#497)
This builds on #423 and completes the coupling of NSE and the simplified-SDC integration. Previously we updated the SDC integrator code path to use the aux data for the thermodynamics. Now we do the test on whether we are in NSE and if so, do a predictor-corrector update of the state to capture the advection + NSE changes.