aprox13 3/∞ (#569)
This change folds the reaction pairs (irmgap, irr1), which did Mg24 -> Al27 -> Si28, and (irsigp, irr1), which did Si28 -> Al27 -> Mg24, into the effective Mg24 <-> Si28 rates. This allows us to remove irr1 from the rate list. More importantly it works toward the goal where every "reaction" known to the network connects an actual pair of species in the network, while any intermediate rates connecting the two can be considered implementation details, which will be helpful for methods looking at microscopic equilibria. This change will be followed by removing the rest of the intermediate proton links.

Note: irmgap, the rate for Mg24(a,p)Al27, is still in the list of rates known to the network. A requirement of the target scheme is that this no longer be true, and it would just be calculated inline while calculating the effective Mg24(a,g)Si28 rate. However it is not straightforward to do this yet -- at the moment, we calculate at once all of the raw (unscreened) rates in the aprox13rat or aprox13 tab routine, and we don't have a way to select a specific raw rate for calculation, which is what would be needed here. So a later change will be to convert the raw rate routines to calculate rates one-by-one (see also #568, which would be needed for calculating the fully screened rates one-by-one), and then we can do that second piece of cleanup.

This causes no harm though; both in the current scheme and the new scheme for iso7, this rate will be calculated but then never explicitly enter into the RHS (only implicitly in the updated Mg24 <-> Si28 rates).