Do not integrate T; get it from e (#496)
We now integrate (X_k, e) instead of (X_k, T) for Strang split coupling.  Closes #471
Note: we always assume a constant volume burn now -- later we can extend this to enthalpy
for constant pressure burns.

To deal with Fortran, our wrapper is going to have to create temporary RHS and Jacobian arrays to pass to actual_rhs that have the temperature component, then strip the temperature component (for the RHS) or convert it to the energy component (for the analytic Jacobian).