kappa_floor should only be applied to Rosseland opacity (#673)
In Castro originally the floor on kappa was only applied to the Rosseland opacity for numerical stability reasons. This change reverts to that previous behavior. This error in the port explains AMReX-Astro/Castro#1539; in that problem setup we had kappa_floor = 1.e-5, which is larger than the actual constant kappa for the Planck mean. We could still have separate floors for kappa_r and kappa_p but that would be done in a separate change.