Merge extra species into species enum, appended at end (#781)
Following up on #779, the "extra" species are now appended to the end of the regular species in various places, including the species enumerator, and the aion and zion arrays. This will make it easy to write loops that do the same thing over the real and extra species by just looping from 1 to NumSpecTotal instead of 1 to NumSpec. Since these are appended at the end, existing code that uses aion and zion in various places will not break, since those loops already explicitly go from 1 to NumSpec. Pasted below is an example of what iso7's enumerator now looks like; the other networks remain unchanged. In this example we have Be8=8, Ca40=9, Ti44=10, NumberTotalSpecies=10, and NumberExtraSpecies=3, as desired.

namespace Species {
   enum NetworkSpecies {
     He4=1,
     C12,
     O16,
     Ne20,
     Mg24,
     Si28,
     Ni56,
     NumberSpecies=Ni56,
     Be8,
     Ca40,
     Ti44,
     NumberExtraSpecies=Ti44-Ni56+1,
     NumberTotalSpecies=Ti44
   };
}