Remove dratdumdy arrays from aprox19 (#785)
The functionality was removed in #686, this just removes the unnecessary dead weight memory.

This changes the Jacobian for aprox19. It turns out that in #567 I forgot to turn off many of the dratdumdy1 terms in aprox19 as I had intended. So this effectively completes what I meant to do at the time.