Remove implicit d/dP terms in aprox19 and aprox21 Jacobians (#840)
In both aprox19 and aprox21 there is the reaction sequence Fe52(a,p)Co55(p,g)Ni56. Co55 is an equilibrium isotope in this reaction chain, and its equilibrium abundance depends on both this reaction sequence as well as the reaction Fe54(p,g)Co55. Since there are two contributions to the effective y(Co55), the rate ends up being somewhat complicated, but the RHS term ends up simply looking like y(Fe52) * y(He4) * y(P) * rate. The Jacobian terms consequently have a d/dP entry. However, although the proton abundance plays a role in determining the rate, the reaction sequence doesn't actually consume or produce a net proton, so it is more appropriate to think of y(P) as being part of the rate rather than part of the RHS. If the y(P) is absorbed into the rate, then this term should not appear in the Jacobian due to the same reason #686 was done: in the upcoming network rewrite, it will not be straightforward (or desirable) to represent any d(rate)/dy term. So this PR proposes removing the term on the basis that it will eventually be removed anyway, and this will make the network rewrite PR easier to evaluate when comparing test_rhs and test_react.