put species at the start of the simplified-SDC vector (#894)
this makes the ordering match that of Strang.  Note that there are some diffs in the output since the jacobian matrix has a different ordering and the linear algebra takes a different sequence of steps now.  But the convergence tests (both reacting and NSE for SDC) show the correct behavior.