NSE solver (#963)
This adds the ability to solve for the NSE state of a network, following the formulation in Calder et al. 2007.  To be solvable, this requires at least 1 nuclei with Z != N (due to the structure of the Jacobian).

This is a first step in allowing nets to automatically transition to NSE.