improve and clean up nse solver (#982)
Added convergence check for hybrj, and switched back to mode 1

Transplant the fine-tuning of initial guess procedure in pynucastro for scipy.fsolve to hybrj solver, making it more robust at solving nse.

change criteria for making the table of solving initial guesses. With fine-tuning, it is easier to solve them.

modify NR procedure to use constraint equations for hybrj (although it is almost never a case we should use nr over hybrj, especially with the fine-tuning procedure added)

Add partition function and spin state to complete it.