Changes to integration/VODE for primordial ISM chemistry (#999)
This PR deals with changes I have made in VODE for primordial chemistry. All the changes basically reflect that since we use number densities for primordial chem, we want to disable some of the sanity checks in VODE that are appropriate for mass fractions. We also do not want to subtract the initial internal energy as is appropriate for nuclear astrochemistry. To disable the checks related to these, I have created two new flags in VODE/_parameters. The default values of these flags is such that users who do not use primordial chem will not need to worry about them, and the code should work just like before. Please let me know what you think.