for Strang + NSE, we need to call EOS (#1098)
after integrating, we call the EOS to get T in case we complete
the burn with NSE, but we were not setting the aux data to be
consistent with the mass fractions, so we got the thermodynamics
wrong.

The reason is that with Castro AMReX-Astro/Castro#2310 we no longer
subtract off the initial energy. This means that when we exit NSE, we need
to have the valid full energy. However, since the burn_t comes in without
the energy initialized, we need to call the EOS with the input rho, T, aux to
get an initial valid e.