swtich Strang over to use T_fixed + update SDC + NSE (#1111)
we were not being consistent with how we handled T_fixed, which
is set if we are doing drive_initial_convection (Castro Strang update
coming soon).  The idea is that the zone's internal energy should
not be computed off of T_fixed, but all of the reaction rates should
be.  This update fixes some issues with NSE and Strang burning where
were we computing e with T_fixed, thus changing the amount of
internal energy in a zone.