add another option to use nse mass fractions for NSE_NET check (#1139)
Since forward and reverse rates depend on the mass fractions, to do a more cruel estimation, we can use nse mass fractions instead. This way, the thermodynamic condition becomes the only deciding factor.

Note here we still need to make sure the current mass fractions are not terribly away from the nse mass fractions, so only allow it after the first check, which checks on whether the current mass fractions are close to the nse mass fractions.