Added flags and updated file paths for building primordial chem test with cmake (#1164)
This PR is related to the cmake functionality for the primordial chem test in Microphysics. It updates the path to Microphysics files in CMakeLists such that they can be found when Microphysics is used as a submodule in another code. It also adds a couple of flags in CMakeLists that the user can specify if they want to build the primordial chem test with amrex.