Compare primordial chem CI output to reference solution (#1214)
In addition to building the primordial chem unit test, we also want to make sure that the comparison to reference solution passes.

The way we test this as follows: we compare a select number of lines from the test output to the stored output. These lines describe the final temperature, internal energy, density, and number densities of electrons, H, H2 and He. If the ratio of the test values to the reference values is larger than 2, the CI will return an exit code 1 and fail the github action.