make self-consistent nse compatible with non-neutron networks. (#1231)
Due to the missing neutron rates, it is difficult to group isotopes together into a single group.

This pr mainly adds another logic of defining a single_group for non-neutron networks. If isotopes fail to form a single group and there is no neutron in the network, then use a second check where we consider nse if isotopes past Z=14(Si) that have odd N or even N form two distinct groups.

Most of the line changes are code formatting.