Make mion and bion inline functions for aprox nets (#1267)
The goal of this change is to remove the global data from actual_network_data.cpp, which in turn allows elimination of that file. Rather than an array, this can be handled with a constexpr function emulating the array. The new function is templated to allow compile-time evaluation of the binding energy. However, not all locations in Microphysics will be readily available to switch to a compile-time evaluation, so a "legacy" interface which accepts the species index at runtime is also provided. (This requires constexpr_for to be available in the network, so it is moved to a separate utility location.)