Convert rprox to the new reactions scheme (#1300)
There were errors in the analytic Jacobian for rprox related to the use of delta1 and delta2 for determining which reaction pathways we were taking. This means results will change; while the new Jacobian is more likely to be correct, it is still not successful for integrating test_react with integrator.jacobian = 1.