don't allow logical runtime parameters (#1385)
when we originally did the Fortran -> C++ conversion, we passed all
of the Fortran logicals into C++ ints.  But now that we are completely
C++, it doesn't make sense to have logical anymore.  There are some
places where we implicitly pass an int to a C++ bool.  This will
help track those down and eliminate things.