remove fast reaction cycle check for NSE_NET check. (#1503)
Fast reaction cycle is one of the check that Kushnir introduced in his paper. However this check only works with neutron present in the network.

In the past, I would just skip this check whenever there is no neutron in the network, which is almost always and it worked fine.

And I've tested with the network that does have neutron without this check, and it gives more or less the same. So I want to get rid of this.

Also changed the indentation to 4 space to be consistent with the rest of the code.