remove the is_jacobian_term_used logic in the linear algebra (#1502)
this was wrong, but turns out, it had no effect because of a bug
in the logic of the is_jacobian_term_used.  That issue is also
fixed here, and the function is kept for use in the future.

note: nets with one species that was not evolved will see a change here.