Optimize fill_merge_indices/nse_grouping for NSE_NET (#1506)
To determine whether we're in NSE for the self-consistent NSE, we need some grouping process. And one of the criteria is to start grouping with the reaction rates that have the fastest (smallest) timescale. However, Kushnir defines the timescale using the molar fraction of the entire group, which complicates things, because the timescale changes after the last successful merge.

I changed it so that we only use the actual molar fraction of the isotope involved in the reaction. This way, we can calculate and store the timescale of all reactions in the beginning and sort the order in which we do the merge by sorting the timescale array.