create a struct to hold the Microphysics runtime parameters (#1433)
This now creates a header extern_type.H with a struct extern_t that holds all
of the runtime parameters for Microphysics. The struct is filled on initialization
and returned by init_extern_parameters(). This can replace the global version
of the runtime parameters.

This also revives the test_parameters unit test to show how to access things.