Use autodiff in screening routines (#1588)
This PR gives the same results for test_screening to within roundoff, and the performance is about the same or slightly better. It also adds a page to the docs about how to use the autodiff library.

One notable change is that the templated networks don't calculate the derivative terms when screening is called from RHS::rhs() (this is also how the pynucastro networks behave). Previously, they would be calculated unnecessarily if integrator.jacobian was set to 1.