updated to replace std::pow with a combo of std::exp and std::log for speedup (#1605)
I replaced most calls to std::pow with a combination of std::exp and std::log to speedup chemistry. Also added free-free cooling, which was missing from the previous version. In the future, I will follow #1586 and #1591 to use fast_exp and fast_log.

The unit test passes, with small differences in the number densities of H and H2 as compared to the reference solution (3.5% and 0.3% respectively). The difference in H is more likely due to free-free cooling.

Co-authored-by: Piyush Sharda <psharda@RSAA-043527.local>