Centralize some runtime parameters and their docs (#1749)
in particular, this moves small_temp and small_dens to the top-level unit_test/_parameters, since every test needs them. We override it in a few tests, since they need much different defaults.

This also moves the documentation about initializing composition up to a new centralized page, so we don't need to repeat it for every test.