Use autodiff in sneut5 (#1799)
This PR intentionally changes as little as possible in terms of numeric operations, so the test results are consistent. We can probably clean up a lot of this later -- there are lots more temporary variables than needed, and sneutf_t stores a bunch of powers of xl mostly used for polynomials, which can use Horner's method instead.