update the ase network (#1813)
Updates the ase network to resemble the original subch_base network. This is slightly different from subch_base where the reverse rates for C12+C12, C12+O16, and O16+O16 are NOT removed.

This is the network that I've been using for testing NSE stuff. This also uses pynucastro/pynucastro#1011 to create DerivedRate for the modified rate.

Also recompute the Q-value, somehow I find it to be more consistent with NSE although it shouldn't matter? Perhaps its just a round-off issue.